requirements.txt
instagrapi
openai
pyphen
wand
python-dotenv
pdfplumber
flask